New approach to traversal
Utilizes vanilla JS to perform traversal instead of babel's traverse utility. Babel's traverse utility is not performant for the types of operations performed by Apollo Aperture. Most of the operations of this devtool are to traverse an application's AST and assemble a hierarchy for D3 to consume and render a React/Apollo component hierarchy. If we were replacing/modifying AST nodes, Babel's solution would be preferred.