Refactor Electron to utilize styled-components
The original implementation of React in Electron used css style sheets. Implementing a CSS loader proved troublesome for electron-webpack's webpack processor, so the application was refactored to utilize styled-components.

D3.js used the d3.json() method to read a file from a URL. This was changed to utilize d3.hierarchy([path-to-data.js]). This was done for two reasons: 1) The new implementation of Electron does not utilize a server and 2) This will allow for async updates to the D3 visualization.