Refactors and updates project so traversal works inside of electron/src/main/traverse
The main entry point is traverseFile.js.
It gets the AST Program Body of the first parameter.
It then gets the first argument to ReactDOM (if it exists). If the file has the ReactDOM then the hierarchy is assembled.
The file anatomy is assembled which includes:
- the default export name and other named imports (including the file path)
- variable declarations

Then the hierarchy of the file is assembled by calling assembleHierarchy.js which recursively calls traverseFile.js if necessary